backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.3
blinker==1.9.0
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cohere==5.14.2
cssselect2==0.8.0
distro==1.9.0
docx2txt==0.9
fastavro==1.10.0
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.57.0
fsspec==2025.3.2
fuzzywuzzy==0.18.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
Levenshtein==0.27.1
lxml==5.3.2
MarkupSafe==3.0.2
openai==1.72.0
packaging==24.2
pdfkit==1.0.0
pillow==11.2.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydyf==0.11.0
pyphen==0.17.2
python-docx==1.1.2
python-Levenshtein==0.27.1
PyYAML==6.0.2
RapidFuzz==3.13.0
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
tinycss2==1.4.0
tinyhtml5==2.0.0
tokenizers==0.21.1
tqdm==4.67.1
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
weasyprint==65.1
webencodings==0.5.1
Werkzeug==3.1.3
zopfli==0.2.3.post1
